+++ Tests/DataTextureAtlasTests.cs
@@ -13,7 +13,7 @@ namespace Tests {
             using var game = TestGame.Create();
             using var texture = new Texture2D(game.GraphicsDevice, 1, 1);
             var atlas = DataTextureAtlas.LoadAtlasData(new TextureRegion(texture), game.RawContent, "Texture.atlas");
-            Assert.AreEqual(atlas.Regions.Count(), 5);
+            Assert.AreEqual(atlas.Regions.Count(), 7);
 
             // no added offset
             var table = atlas["LongTableUp"];
@@ -21,7 +21,7 @@ namespace Tests {
             Assert.AreEqual(table.PivotPixels, new Vector2(16, 48 - 32));
 
             // added offset
-            var table2 = atlas["LongTableRight"];
+            var table2 = atlas["LongTableLeft"];
             Assert.AreEqual(table2.Area, new Rectangle(64, 32, 64, 48));
             Assert.AreEqual(table2.PivotPixels, new Vector2(112 - 64, 48 - 32));
         }